Support multiple BIOS versions
Rework project layout to support multiple versions of the BIOS, using
makepp's repository feature to mount source files into separate
directories for each BIOS version.

Also update to the latest version of EuroASM so we can run the native
Linux version of the assembler.  This does mean we need to patch the
EuroASM source and build a local copy, because by default it does not
consider symbolic links to be valid source files.  The makefiles take
care of all of this for us.